+++ bot/src/events/ready/ready.js
@@ -6,5 +6,11 @@ module.exports = class ReadyEvent extends BaseEvent {
   }
   async run (client) {
     console.log(client.user.tag + ' has logged in.');
+    client.user.setPresence({
+      activity: {
+        name: 'test',
+        status: 0,
+      }
+    })
   }
 }